🎨 feat: 장바구니 선택삭제 zustand사용하여 리팩토링

+++ src/data/actions/delete_all_cart_action.ts
@@ -1,28 +1,27 @@
 'use server';
 
 export async function deleteAllCartAction(
-  checkedItems: number[],
-  token: string | undefined,
+  _: { status?: boolean; error: string },
+  formData: FormData,
 ) {
-  // const stringItems = formData.get('checkedItems'); //배열 데이터를 넘겨받을때는 getAll을 사용
-  // const stringItems = formData.get('checkedItems'); //배열 데이터를 넘겨받을때는 getAll을 사용
-  // const token = formData.get('token')?.toString();
+  const stringItems = formData.get('checkedItems'); //배열 데이터를 넘겨받을때는 getAll을 사용
+  const token = formData.get('token')?.toString();
 
-  // const checkedItems: number[] = stringItems?.map(item => Number(item));
-  // console.log('ㅎㅇㅎㅇ4', checkedItems);
-  // if (!checkedItems) {
-  console.log('체크 아이템즈', checkedItems);
-  if (!checkedItems) {
-    // return {
-    //   status: false,
-    //   error: '삭제할 상품이 없습니다.',
-    // };
+  console.log('ㅎㅇㅎㅇ3', stringItems);
+
+  //조건식 이렇게 하면 parse에 타입에러 없어짐
+  if (!stringItems || typeof stringItems !== 'string') {
+    return {
+      status: false,
+      error: '삭제할 상품이 없습니다.',
+    };
   }
+  const deleteList = JSON.parse(stringItems);
   try {
     const response = await fetch(`${process.env.NEXT_PUBLIC_API_URL}/carts/`, {
       method: 'DELETE',
       body: JSON.stringify({
-        carts: checkedItems,
+        carts: deleteList,
       }),
       headers: {
         Authorization: `Bearer ${token}`,
@@ -33,16 +32,15 @@ export async function deleteAllCartAction(
     if (!response.ok) {
       throw new Error(response.statusText);
     }
-    //재검증
-    // return {
-    //   status: true,
-    //   error: '',
-    // };
+
+    return {
+      status: true,
+      error: '',
+    };
   } catch (err) {
-    // return {
-    //   status: false,
-    //   error: `상품 삭제에 실패했습니다: ${err}`,
-    // };
-    console.error(err);
+    return {
+      status: false,
+      error: `상품 삭제에 실패했습니다: ${err}`,
+    };
   }
 }
